test: move Rust tests into tests directories
